+++ src/components/Admin/ClubAdmin.tsx
@@ -1,4 +1,5 @@
-import { useQuery } from '@apollo/client'
+import { useQuery, useSubscription } from '@apollo/client'
+import log from '@kengoldfarb/log'
 import {
 	createStyles,
 	Container,
@@ -7,16 +8,23 @@ import {
 	Space,
 	Center,
 	Loader,
-	Divider
+	Divider,
+	Button
 } from '@mantine/core'
 import { showNotification } from '@mantine/notifications'
-import { useWallet } from '@meemproject/react'
-import Cookies from 'js-cookie'
+import { makeFetcher, MeemAPI } from '@meemproject/api'
+import { diamondABI } from '@meemproject/meem-contracts'
+import { LoginState, useWallet } from '@meemproject/react'
+import { ethers } from 'ethers'
 import { useRouter } from 'next/router'
 import React, { useEffect, useState } from 'react'
 import { ArrowLeft, Check } from 'tabler-icons-react'
-import { GetClubQuery, MeemContracts } from '../../../generated/graphql'
-import { GET_CLUB } from '../../graphql/clubs'
+import {
+	GetBundleByIdQuery,
+	GetClubSubscriptionSubscription,
+	MeemContracts
+} from '../../../generated/graphql'
+import { GET_BUNDLE_BY_ID, SUB_CLUB } from '../../graphql/clubs'
 import clubFromMeemContract, { Club } from '../../model/club/club'
 import { ClubAdminDappSettingsComponent } from './ClubAdminDappsSettings'
 import { ClubAdminMembershipSettingsComponent } from './ClubAdminMembershipSettings'
@@ -115,6 +123,13 @@ const useStyles = createStyles(theme => ({
 			borderColor: 'transparent'
 		}
 	},
+	buttonCreate: {
+		backgroundColor: 'black',
+		'&:hover': {
+			backgroundColor: theme.colors.gray[8]
+		},
+		borderRadius: 24
+	},
 	tabs: {
 		display: 'flex',
 		flexDirection: 'row'
@@ -188,6 +203,16 @@ export const ClubAdminComponent: React.FC<IProps> = ({ slug }) => {
 	const wallet = useWallet()
 
 	const [currentTab, setCurrentTab] = useState<Tab>(Tab.Membership)
+	const [isCreatingSafe, setIsCreatingSafe] = useState(false)
+	const [bundleFunctionSelectors, setBundleFunctionSelectors] = useState<
+		string[]
+	>([])
+	const [clubFunctionSelectors, setClubFunctionSelectors] = useState<
+		string[]
+	>([])
+	const [bundleABI, setBundleABI] = useState<string>()
+	const [shouldShowUpgrade, setShouldShowUpgrade] = useState(false)
+	const [isUpgradingClub, setIsUpgradingClub] = useState(false)
 
 	const navigateToClubDetail = () => {
 		router.push({ pathname: `/${slug}` })
@@ -209,17 +234,26 @@ export const ClubAdminComponent: React.FC<IProps> = ({ slug }) => {
 		loading,
 		error,
 		data: clubData
-	} = useQuery<GetClubQuery>(GET_CLUB, {
+	} = useSubscription<GetClubSubscriptionSubscription>(SUB_CLUB, {
 		variables: { slug }
 	})
+
+	const { data: bundleData } = useQuery<GetBundleByIdQuery>(
+		GET_BUNDLE_BY_ID,
+		{
+			variables: {
+				id: process.env.NEXT_PUBLIC_MEEM_BUNDLE_ID
+			}
+		}
+	)
+
 	const [isLoadingClub, setIsLoadingClub] = useState(true)
 	const [club, setClub] = useState<Club>()
 
 	useEffect(() => {
 		if (
 			// Note: walletContext thinks logged in = LoginState.unknown, using cookies here
-			Cookies.get('meemJwtToken') === undefined ||
-			Cookies.get('walletAddress') === undefined
+			wallet.loginState === LoginState.NotLoggedIn
 		) {
 			router.push({
 				pathname: '/authenticate',
@@ -228,10 +262,10 @@ export const ClubAdminComponent: React.FC<IProps> = ({ slug }) => {
 				}
 			})
 		}
-	}, [router, slug])
+	}, [router, slug, wallet.loginState])
 
 	useEffect(() => {
-		async function getClub(data: GetClubQuery) {
+		async function getClub(data: GetClubSubscriptionSubscription) {
 			const possibleClub = await clubFromMeemContract(
 				wallet,
 				wallet.isConnected ? wallet.accounts[0] : '',
@@ -256,13 +290,103 @@ export const ClubAdminComponent: React.FC<IProps> = ({ slug }) => {
 		wallet.isConnected
 	])
 
+	// const refetchClub = async () => {
+	// 	log.debug('refetching club...')
+	// 	const newClub = await refetch()
+	// 	if (newClub.data.MeemContracts.length > 0) {
+	// 		const possibleClub = await clubFromMeemContract(
+	// 			wallet,
+	// 			wallet.isConnected ? wallet.accounts[0] : '',
+	// 			newClub.data.MeemContracts[0] as MeemContracts
+	// 		)
+
+	// 		if (possibleClub && possibleClub.name) {
+	// 			log.debug('got new club data, setting it...')
+	// 			setClub(possibleClub)
+	// 		}
+	// 	}
+	// }
+
+	useEffect(() => {
+		if (bundleData?.Bundles[0]) {
+			let allFunctionSelectors: string[] = []
+			bundleData.Bundles[0].BundleContracts.forEach(bc => {
+				allFunctionSelectors = [
+					...allFunctionSelectors,
+					...bc.functionSelectors
+				]
+			})
+
+			setBundleFunctionSelectors(allFunctionSelectors)
+			setBundleABI(bundleData.Bundles[0].abi)
+		}
+	}, [bundleData])
+
+	useEffect(() => {
+		const fetchFacets = async () => {
+			if (bundleABI && club?.address) {
+				try {
+					const diamond = new ethers.Contract(
+						club.address,
+						diamondABI,
+						wallet.signer
+					)
+
+					const facets = await diamond.facets()
+					let currentSelectors: string[] = []
+					facets.forEach(
+						(facet: { target: string; selectors: string[] }) => {
+							currentSelectors = [
+								...currentSelectors,
+								...facet.selectors
+							]
+						}
+					)
+					setClubFunctionSelectors(currentSelectors)
+				} catch (e) {
+					log.crit(e)
+				}
+			}
+		}
+		fetchFacets()
+	}, [bundleABI, club, wallet])
+
+	useEffect(() => {
+		if (
+			bundleFunctionSelectors.length > 0 &&
+			clubFunctionSelectors.length > 0
+		) {
+			for (let i = 0; i < bundleFunctionSelectors.length; i += 1) {
+				const clubFS = clubFunctionSelectors.find(
+					fs => fs === bundleFunctionSelectors[i]
+				)
+				if (!clubFS) {
+					setShouldShowUpgrade(true)
+					return
+				}
+			}
+
+			// If we made it this far, all selectors are accounted for and the upgrade was successful
+			if (shouldShowUpgrade) {
+				showNotification({
+					title: 'Club Upgraded!',
+					color: 'green',
+					icon: <Check />,
+					message: `The club has been upgraded to the latest version.`
+				})
+				setShouldShowUpgrade(false)
+				setIsUpgradingClub(false)
+			}
+		}
+	}, [bundleFunctionSelectors, clubFunctionSelectors, shouldShowUpgrade])
+
 	return (
 		<>
 			{isLoadingClub && (
 				<Container>
 					<Space h={120} />
 					<Center>
-						<Loader />
+						<Loader color="red" variant="bars" />
 					</Center>
 				</Container>
 			)}
@@ -286,6 +410,8 @@ export const ClubAdminComponent: React.FC<IProps> = ({ slug }) => {
 								/>
 							</a>
 							<Image
+								width={80}
+								height={80}
 								className={classes.clubLogoImage}
 								src={club.image}
 							/>
@@ -307,6 +433,7 @@ export const ClubAdminComponent: React.FC<IProps> = ({ slug }) => {
 												`${window.location.origin}/${club.slug}`
 											)
 											showNotification({
+												radius: 'lg',
 												title: 'Club URL copied',
 												autoClose: 2000,
 												color: 'green',
@@ -402,6 +529,7 @@ export const ClubAdminComponent: React.FC<IProps> = ({ slug }) => {
 												club.address ?? ''
 											)
 											showNotification({
+												radius: 'lg',
 												title: 'Address copied',
 												autoClose: 2000,
 												color: 'green',
@@ -413,6 +541,212 @@ export const ClubAdminComponent: React.FC<IProps> = ({ slug }) => {
 										width={20}
 									/>
 								</div>
+
+								<Space h={8} />
+
+								<Text>{`You can use this address to token-gate third-party apps and tools, such as creating an exclusive Discord community with Collab.Land. Every club member holds this club's token.`}</Text>
+
+								{shouldShowUpgrade && (
+									<>
+										<Space h={'xl'} />
+										<Divider />
+										<Space h={'xl'} />
+										<Text
+											className={
+												classes.clubIntegrationsSectionTitle
+											}
+										>
+											Upgrade Club Contract
+										</Text>
+										<div
+											className={
+												classes.contractAddressContainer
+											}
+										>
+											<div>
+												<Text>
+													A new version of Clubs is
+													available! Upgrade to take
+													advantage of all the new
+													features.
+												</Text>
+											</div>
+										</div>
+										<Space h={'xs'} />
+										<Button
+											loading={isUpgradingClub}
+											disabled={isUpgradingClub}
+											className={classes.buttonCreate}
+											onClick={async () => {
+												try {
+													if (!club?.id) {
+														return
+													}
+													setIsUpgradingClub(true)
+													const upgradeClubFetcher =
+														makeFetcher<
+															MeemAPI.v1.UpgradeClub.IQueryParams,
+															MeemAPI.v1.UpgradeClub.IRequestBody,
+															MeemAPI.v1.UpgradeClub.IResponseBody
+														>({
+															method: MeemAPI.v1
+																.UpgradeClub
+																.method
+														})
+
+													await upgradeClubFetcher(
+														MeemAPI.v1.UpgradeClub.path(
+															{
+																meemContractId:
+																	club.id
+															}
+														)
+													)
+												} catch (e) {
+													log.crit(e)
+													showNotification({
+														title: 'Error Upgrading Club',
+														color: 'red',
+														message: `Something went wrong during the upgrade.`
+													})
+													setIsUpgradingClub(false)
+												}
+											}}
+										>
+											Upgrade Contract
+										</Button>
+									</>
+								)}
+
+								<Space h={'xl'} />
+								<Divider />
+								<Space h={'xl'} />
+								<Text
+									className={
+										classes.clubIntegrationsSectionTitle
+									}
+								>
+									Club Treasury Address
+								</Text>
+								{club.gnosisSafeAddress && (
+									<>
+										<div
+											className={
+												classes.contractAddressContainer
+											}
+										>
+											<Text
+												className={
+													classes.clubContractAddress
+												}
+											>
+												{club.gnosisSafeAddress}
+											</Text>
+											<Image
+												className={classes.copy}
+												src="/copy.png"
+												height={20}
+												onClick={() => {
+													navigator.clipboard.writeText(
+														club.gnosisSafeAddress ??
+															''
+													)
+													showNotification({
+														radius: 'lg',
+														title: 'Address copied',
+														autoClose: 2000,
+														color: 'green',
+														icon: <Check />,
+
+														message: `This club's treasury address was copied to your clipboard.`
+													})
+												}}
+												width={20}
+											/>
+										</div>
+										<Space h={8} />
+
+										<Text>{`Your club's treasury was set up when the club was created. You can manage your treasury (including signing transactions and adding members) using the button below.`}</Text>
+										<Space h={'xs'} />
+
+										<Button
+											className={classes.buttonCreate}
+											onClick={() => {
+												window.open(
+													`https://gnosis-safe.io/app/${
+														process.env
+															.NEXT_PUBLIC_CHAIN_ID ===
+														'4'
+															? 'rin'
+															: 'matic'
+													}:${
+														club.gnosisSafeAddress
+													}/home`
+												)
+											}}
+										>
+											View Treasury
+										</Button>
+									</>
+								)}
+
+								{!club.gnosisSafeAddress && (
+									<Button
+										className={classes.buttonCreate}
+										disabled={isCreatingSafe}
+										loading={isCreatingSafe}
+										onClick={async () => {
+											if (!club.id || !club.admins) {
+												return
+											}
+											try {
+												setIsCreatingSafe(true)
+												const createSafeFetcher =
+													makeFetcher<
+														MeemAPI.v1.CreateClubSafe.IQueryParams,
+														MeemAPI.v1.CreateClubSafe.IRequestBody,
+														MeemAPI.v1.CreateClubSafe.IResponseBody
+													>({
+														method: MeemAPI.v1
+															.CreateClubSafe
+															.method
+													})
+
+												await createSafeFetcher(
+													MeemAPI.v1.CreateClubSafe.path(
+														{
+															meemContractId:
+																club.id
+														}
+													),
+													undefined,
+													{
+														safeOwners: club.admins
+													}
+												)
+												await new Promise(f =>
+													setTimeout(f, 10000)
+												)
+
+												// refetchClub()
+
+												setIsCreatingSafe(false)
+											} catch (e) {
+												log.crit(e)
+												setIsCreatingSafe(false)
+												showNotification({
+													radius: 'lg',
+													title: 'Wallet creation failed.',
+													message:
+														'We were unable to create treasury for your club. Please refresh the page and try again.',
+													color: 'red'
+												})
+											}
+										}}
+									>
+										Create Treasury
+									</Button>
+								)}
 
 								<Space h={'xl'} />
 								<Divider />