+++ src/components/Admin/ClubAdminDappsSettings.tsx
@@ -360,6 +360,7 @@ export const ClubAdminDappSettingsComponent: React.FC<IProps> = ({ club }) => {
 				new URL(currentIntegrationUrl)
 			} catch (_) {
 				showNotification({
+					radius: 'lg',
 					title: 'Oops!',
 					message: 'Please enter a valid URL for this integration.'
 				})
@@ -400,6 +401,7 @@ export const ClubAdminDappSettingsComponent: React.FC<IProps> = ({ club }) => {
 				log.debug(e)
 				setIsSavingChanges(false)
 				showNotification({
+					radius: 'lg',
 					title: 'Oops!',
 					message:
 						'Unable to save this integration. Please get in touch!'
@@ -677,7 +679,7 @@ export const ClubAdminDappSettingsComponent: React.FC<IProps> = ({ club }) => {
 				)}
 				{loading && !inteData && (
 					<>
-						<Loader />
+						<Loader color="red" variant="bars" />
 					</>
 				)}
 				{!loading && error && (
@@ -717,6 +719,7 @@ export const ClubAdminDappSettingsComponent: React.FC<IProps> = ({ club }) => {
 				/>
 				<Modal
 					centered
+					overlayBlur={8}
 					closeOnClickOutside={false}
 					closeOnEscape={false}
 					radius={16}