+++ src/components/Admin/IntegrationModals/ClubAdminVerifyTwitterModal.tsx
@@ -132,6 +132,7 @@ export const ClubAdminVerifyTwitterModal: React.FC<IProps> = ({
 				})
 			log.debug(body)
 			showNotification({
+				radius: 'lg',
 				title: 'Success!',
 				autoClose: 5000,
 				color: 'green',
@@ -144,6 +145,7 @@ export const ClubAdminVerifyTwitterModal: React.FC<IProps> = ({
 		} catch (e) {
 			log.debug(e)
 			showNotification({
+				radius: 'lg',
 				title: 'Verification failed',
 				autoClose: 5000,
 				color: 'red',
@@ -162,6 +164,7 @@ export const ClubAdminVerifyTwitterModal: React.FC<IProps> = ({
 				closeOnEscape={false}
 				withCloseButton={step !== Step.Verifying}
 				radius={16}
+				overlayBlur={8}
 				padding={'sm'}
 				opened={isOpened}
 				title={